+++ lib/view/widgets/navigation_drawer.dart
@@ -32,7 +32,7 @@ class BMDrawer extends StatelessWidget {
           ),
           ListTile(
             //Add Icons from assets
-            leading: Icon(Icons.draw_outlined),
+            leading: const Icon(Icons.draw_outlined),
             title: const Text('Draw Clipart'),
             onTap: () {
               Navigator.pushNamed(context, '/drawBadge');
