+++ lib/virtualbadge/widgets/badge_widget.dart
